requirements.txt
replicate
python-dotenv